Improved color refinements
