+++ commerce-cloud-code/app_yourshophere/cartridge/utils/lazyload.js
@@ -1,8 +1,8 @@
 module.exports = function addLazyProperty(object, name, callback) {
     Object.defineProperty(object, name, {
-        get: function () {
+        get() {
             return callback.apply(object, [object]);
         },
-        configurable: true
+        configurable: true,
     });
-}
+};
